+++ web/.storybook/webpack.config.js
@@ -1,0 +1,6 @@
+const path = require("path");
+
+module.exports = ({ config }) => {
+  config.resolve.alias["@"] = path.resolve(__dirname, "..", "src");
+  return config;
+};
