+++ web/src/components/card/index.stories.tsx
@@ -14,10 +14,14 @@ type Story = StoryObj<typeof Component>;
 
 // More on writing stories with args: https://storybook.js.org/docs/7.0/react/writing-stories/args
 export const Card: Story = {
-  render: (args) => (
+  render: (args: any) => (
     <div className="w-[380px]">
       <Component {...args} />
     </div>
   ),
-  args: {},
+  args: {
+    title: "Card Title",
+    address: "1234 Main St, City, State, 12345",
+    imageUrls: ["https://picsum.photos/200"],
+  },
 };
