+++ web/src/hooks/useAuth.ts
@@ -25,7 +25,7 @@ export const useAuth = () => {
     if (user) {
       addUser(JSON.parse(user));
     }
-  }, []);
+  }, [addUser, getItem]);
 
   const signup = async (data: SignUpParams) => {
     if (
